% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linux-binaries.R
\name{preinstall_linux_binaries}
\alias{preinstall_linux_binaries}
\title{The goal of this function is to speed up installation of dependencies.
This is done in two ways:
\itemize{
\item Try to get precompiled binaries when available (mainly on ubuntu)
\item Download files in parallel using curl
The latter can be removed once fixed in base-R: https://github.com/r-devel/r-svn/pull/155
}}
\usage{
preinstall_linux_binaries(packages)
}
\description{
The goal of this function is to speed up installation of dependencies.
This is done in two ways:
\itemize{
\item Try to get precompiled binaries when available (mainly on ubuntu)
\item Download files in parallel using curl
The latter can be removed once fixed in base-R: https://github.com/r-devel/r-svn/pull/155
}
}
