% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recheck.R
\name{revdep_check}
\alias{revdep_check}
\title{Reverse Dependency Check}
\usage{
revdep_check(
  sourcepkg,
  which = "strong",
  repos = "https://cloud.r-project.org"
)
}
\arguments{
\item{sourcepkg}{path to a source package tarball}

\item{which}{passed to \code{tools::package_dependencies}; set to "most" to
also check reverse suggests.}

\item{repos}{vector of repos to find reverse dependencies}
}
\description{
Run a reverse dependency check similar to CRAN.
}
