% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recheck.R
\name{revdep_check}
\alias{revdep_check}
\title{Reverse Dependency Check}
\usage{
revdep_check(sourcepkg, which = "strong")
}
\arguments{
\item{sourcepkg}{path to a source package tarball}

\item{which}{passed to \code{tools::check_packages_in_dir}}
}
\description{
Run a reverse dependency check similar to CRAN.
}
