% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recheck.R
\name{revdep_check}
\alias{revdep_check}
\title{Reverse Dependency Check}
\usage{
revdep_check(pkg, checkdir = "recheck", which = "strong", recursive = FALSE)
}
\arguments{
\item{pkg}{name of a package}

\item{checkdir}{directory where revdeps are downloaded and checked.}

\item{which}{passed to \code{tools::package_dependencies}}

\item{recursive}{passed to \code{tools::package_dependencies}}
}
\description{
Run checks on known dependents of a package.
}
